cloud/requirements.txt
functions-framework==3.*
google-api-core==2.15.0
google-auth==2.26.2
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
mysqlclient==2.2.1
numpy==1.26.2
pillow==10.2.0
python-dotenv==1.0.0
requests==2.31.0
face-recognition==1.3.0       
face-recognition-models==0.3.0
openai==1.7.2